Fixing some cli stuff and the event perimeter id bug. Parallelizing spatial merging, and implementing new overall technique for merging in time and space. Also ignoring 0 and -1s when finding window events in temporal range
