+++ environment.yml
@@ -1,0 +1,26 @@
+name: firedpy  # Replace with your desired environment name
+channels:
+  - conda-forge  # Use the conda-forge channel for some packages
+dependencies:
+  - beautifulsoup4
+  - dask
+  - descartes
+  - geopandas
+  - fiona
+  - gdal
+  - lxml
+  - matplotlib
+  - netcdf4
+  - numpy
+  - pandas
+  - pycurl
+  - pyyaml
+  - rasterio
+  - requests
+  - toolz
+  - tqdm
+  - xarray
+  - paramiko
+  - pytest-shutil
+  - openblas
+  - scipy
