+++ utilities/create_readme.py
@@ -51,9 +51,9 @@ def makeReadMe(proj_dir, tilename, file_base, input, first_date, last_date, ecor
         print("ABSTRACT\n", file=text_file)
         print("-------------------\n", file=text_file)
         if input == 3:
-            print("This is event-level polygons for the fire event delineation (FIRED) product for MODIS grid tiles", *name , "from {}  to {}. It is derived from the MODIS MCD64A1 burned area product (see https://lpdaac.usgs.gov/products/mcd64a1v006/ for more details). The MCD64A1 is a monthly raster grid of estimated burned dates. Firedpy (www.github.com/earthlab/firedpy) is an algorithm that converts these rasters into events by stacking the entire time series into a spatial-temporal data cube, then uses an algorithm to assign event identification numbers to pixels that fit into the same 3-dimensional spatial temporal window. This particular dataset was created using a spatial parameter of {} pixels and {} days. If daily polygons are included, the event identification numbers are the same for both files, but the event-level product has only single polygons for each entire event, while the daily product has separate polygons for each date per event. See the associated paper for more details on the methods and more:\n".format(first_event, last_event, spatial_param, temporal_param), file=text_file)
+            print("This is event-level polygons for the fire event delineation (FIRED) product for MODIS grid tiles", *name , "from {}  to {}. It is derived from the MODIS MCD64A1 burned area product (see https://lpdaac.usgs.gov/products/mcd64a1v006/ for more details). The MCD64A1 is a monthly raster grid of estimated burned dates. Firedpy (www.github.com/earthlab/src) is an algorithm that converts these rasters into events by stacking the entire time series into a spatial-temporal data cube, then uses an algorithm to assign event identification numbers to pixels that fit into the same 3-dimensional spatial temporal window. This particular dataset was created using a spatial parameter of {} pixels and {} days. If daily polygons are included, the event identification numbers are the same for both files, but the event-level product has only single polygons for each entire event, while the daily product has separate polygons for each date per event. See the associated paper for more details on the methods and more:\n".format(first_event, last_event, spatial_param, temporal_param), file=text_file)
         else:
-            print("This is event-level polygons for the fire event delineation (FIRED) product for", *name , "from {}  to {}. It is derived from the MODIS MCD64A1 burned area product (see https://lpdaac.usgs.gov/products/mcd64a1v006/ for more details). The MCD64A1 is a monthly raster grid of estimated burned dates. Firedpy (www.github.com/earthlab/firedpy) is an algorithm that converts these rasters into events by stacking the entire time series into a spatial-temporal data cube, then uses an algorithm to assign event identification numbers to pixels that fit into the same 3-dimensional spatial temporal window. This particular dataset was created using a spatial parameter of {} pixels and {} days. If daily polygons are included, the event identification numbers are the same for both files, but the event-level product has only single polygons for each entire event, while the daily product has separate polygons for each date per event. See the associated paper for more details on the methods and more:\n".format(first_event, last_event, spatial_param, temporal_param), file=text_file)
+            print("This is event-level polygons for the fire event delineation (FIRED) product for", *name , "from {}  to {}. It is derived from the MODIS MCD64A1 burned area product (see https://lpdaac.usgs.gov/products/mcd64a1v006/ for more details). The MCD64A1 is a monthly raster grid of estimated burned dates. Firedpy (www.github.com/earthlab/src) is an algorithm that converts these rasters into events by stacking the entire time series into a spatial-temporal data cube, then uses an algorithm to assign event identification numbers to pixels that fit into the same 3-dimensional spatial temporal window. This particular dataset was created using a spatial parameter of {} pixels and {} days. If daily polygons are included, the event identification numbers are the same for both files, but the event-level product has only single polygons for each entire event, while the daily product has separate polygons for each date per event. See the associated paper for more details on the methods and more:\n".format(first_event, last_event, spatial_param, temporal_param), file=text_file)
         print("Balch, J.K.; St. Denis, L.A.; Mahood, A.L.; Mietkiewicz, N.P.; Williams, T.M.; McGlinchy, J.; Cook, M.C. FIRED (Fire Events Delineation): An Open, Flexible Algorithm and Database of US Fire Events Derived from the MODIS Burned Area Product (2001–2019). Remote Sens. 2020, 12, 3498. https://doi.org/10.3390/rs12213498 \n", file=text_file)
         print("""-------------------\n
 GENERAL INFORMATION\n
